+++ windows/flutter/generated_plugins.cmake
@@ -3,8 +3,10 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  fullscreen_window
   just_audio_windows
   url_launcher_windows
+  webview_win_floating
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST